pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-blanket"
version = "0.1.0"
description = "Unified Python library for LLM APIs (OpenAI, Anthropic, Gemini, xAI, Groq, custom)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Yoseph Berhanu Alebachew" }]
keywords = ["llm", "openai", "anthropic", "gemini", "groq", "xai", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "httpx>=0.25",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.39"]
gemini = ["google-genai>=1.0"]
all = [
    "llm-blanket[openai,anthropic,gemini]",
]

[project.urls]
Repository = "https://github.com/yosephberhanu/llm-blanket"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
